feat: added email confirmation code on backend

+++ backend/src/controllers/user/deleteAccount.js
@@ -1,10 +1,14 @@
 import jwt from 'jsonwebtoken';
 import * as db from '../../db/index.js';
+import { validateCode } from '../../utils/validateCode.js';
 
 export async function deleteAccount(req, res) {
     try {
         const token = req.headers.authorization.split(' ')[1];
         const { email } = jwt.decode(token);
+        if (!(await validateCode(req, email))) {
+            return res.status(401).json({ error: 'Invalid code.' });
+        }
         await db.deleteAccount(email);
         return res.status(200).json({ message: 'User deleted successfully.' });
     } catch (err) {
